% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.unique.R
\name{group.unique}
\alias{group.unique}
\title{Count the unique values of a variable per a grouped variable.}
\usage{
group.unique(dataset, grouped_variable, unique_variable)
}
\arguments{
\item{dataset}{The dataset containing the variables that you wish to count unique values of a group from}

\item{grouped_variable}{Name of the column containing the variable that you want to group.}

\item{unique_variable}{Name of the column containing the variable that you want to extract the number of unique/distinct values from for each group in the grouped_variable.}
}
\value{
A tibble with two columns. The first column is the grouped values with the column name of the grouped_variable. The second column is the number of unique values of the other variable (unique_variable) per each group, with the column name "count_of_unique".
}
\description{
This function counts the unique (aka distinct) values of a variable that exist in each group of another variable of a dataset. It will produce a tibble with two columns. The first is the grouped values, and the second is the number of unique values per group.
}
\examples{
library(datateachr)
group.unique(vancouver_trees, neighbourhood_name, common_name)
library(gapminder)
group.unique(gapminder, year, country)
}
